added new table(model, dto), repository, service and controller

+++ TravelCompanyClient/Models/CityDTO.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int? CountryId { get; set; }
-        public int? ClimateId { get; set; }
+        public byte? ClimateId { get; set; }
         public string Name { get; set; }
     }
 }
